% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPZone_setup.R
\name{check_setup}
\alias{check_setup}
\title{Checks if the setup function has been run, or can be run automatically. If so, HPZone_setup() is called. If not, an error is thrown.}
\usage{
check_setup()
}
\value{
TRUE if setup has been run or can be run by this function.
}
\description{
Checks if the setup function has been run, or can be run automatically. If so, HPZone_setup() is called. If not, an error is thrown.
}
\examples{
check_setup()
}
