% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HPZone_fields}
\alias{HPZone_fields}
\title{Endpoints and fields available in the HPZone API}
\format{
A tibble containing field names and metadata.
}
\usage{
HPZone_fields
}
\description{
Endpoints and fields available in the HPZone API
}
\keyword{datasets}
