% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPZone_make_valid.R
\name{HPZone_make_valid}
\alias{HPZone_make_valid}
\title{Takes a list of fields or endpoints and corrects them. This allows for short hand usage without having to check the documentation, e.g. 'Date of onset' instead of 'Date_of_onset'}
\usage{
HPZone_make_valid(endpoints = NULL, fields = NULL)
}
\arguments{
\item{endpoints}{A list of desired endpoints.}

\item{fields}{A list of desired fields.}
}
\value{
A list of properly formatted fields or endpoints.
}
\description{
Takes a list of fields or endpoints and corrects them. This allows for short hand usage without having to check the documentation, e.g. 'Date of onset' instead of 'Date_of_onset'
}
