% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPZone_setup.R
\name{HPZone_store_credentials}
\alias{HPZone_store_credentials}
\title{Safely stores the HPZone API details, so you don't need to put them in your script. The default OS keyring backend is used through the keyring-package. Values are stored in an encrypted format to prevent harvesting by other applications.
Note that the actual values are not supplied as arguments, but requested using rstudioapi password prompts.}
\usage{
HPZone_store_credentials(client_id = T, client_secret = T)
}
\arguments{
\item{client_id}{True / false: whether the client_id should be stored.}

\item{client_secret}{True / false: whether the client_secret should be stored.}
}
\description{
Safely stores the HPZone API details, so you don't need to put them in your script. The default OS keyring backend is used through the keyring-package. Values are stored in an encrypted format to prevent harvesting by other applications.
Note that the actual values are not supplied as arguments, but requested using rstudioapi password prompts.
}
\examples{
# simply execute this line to store credentials
HPZone_store_credentials()
# after use, setup can be ran without arguments:
HPZone_setup()
}
