% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_HPZone_token.R
\name{test_HPZone_token}
\alias{test_HPZone_token}
\title{Tests the supplied HPZone token against the token endpoint. Note that this does not check the type of access, only the functionality of the token.}
\usage{
test_HPZone_token()
}
\value{
No return value.
}
\description{
Tests the supplied HPZone token against the token endpoint. Note that this does not check the type of access, only the functionality of the token.
}
\examples{
HPZone_setup()
# This will print the results.
test_HPZone_token()
}
