% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPZone_convert_dates.R
\name{HPZone_convert_dates}
\alias{HPZone_convert_dates}
\title{Converts all relevant columns in the dataset to Date types.}
\usage{
HPZone_convert_dates(data, search = "dat(e|um)")
}
\arguments{
\item{data}{A dataset returned by the API.}

\item{search}{Column names to search for. Default: anything containing 'date' or 'datum'.}
}
\value{
An equivalent data.frame with Date types in the correct columns.
}
\description{
Converts all relevant columns in the dataset to Date types.
}
