% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPZone_request.R
\name{HPZone_request_paginated}
\alias{HPZone_request_paginated}
\title{Performs a HPZone request with the given parameters.}
\usage{
HPZone_request_paginated(query, n_max = 500, scope = API_env$scope_standard)
}
\arguments{
\item{query}{A GraphQL query to send to the HPZone API. Note that keywords 'skip' and 'take' cannot be present in the query; this function will add them.}

\item{n_max}{Maximum number of entries to request per call.}

\item{scope}{The desired scope; either standard or extended.}
}
\value{
A data frame containing all the responses gathered from the API. Only the items will be returned.
}
\description{
Performs a HPZone request with the given parameters.
}
