% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPZone_setup.R
\name{HPZone_setup}
\alias{HPZone_setup}
\title{Initialisation function to define API credentials. This function must be called before anything else, as the details supplied in this call are required for use.}
\usage{
HPZone_setup(
  client_id,
  client_secret,
  standard,
  extended,
  token_url = "https://login.ggdghor.nl/infactdev/oauth2/v1/token",
  data_url = "http://85.90.69.239:8899/Edie"
)
}
\arguments{
\item{data_url}{}
}
\description{
Initialisation function to define API credentials. This function must be called before anything else, as the details supplied in this call are required for use.
}
