% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPZoneAPI-package.R
\docType{package}
\name{HPZoneAPI-package}
\alias{HPZoneAPI}
\alias{HPZoneAPI-package}
\title{HPZoneAPI: HPZone API Interface in R}
\description{
Package that simplifies the use of HPZone's API. Most of the annoying and labor-intensive parts of the interface are handled by wrapper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggdatascience/HPZoneAPI}
  \item Report bugs at \url{https://github.com/ggdatascience/HPZoneAPI/issues}
}

}
\author{
\strong{Maintainer}: Dijkstra Aart \email{a.dijkstra@ggdnog.nl} [copyright holder]

}
\keyword{internal}
